pyproject.toml
[project]
name = "cloudone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-mgmt-advisor>=9.0.0",
    "azure-mgmt-carbonoptimization>=1.0.0",
    "azure-mgmt-resource>=24.0.0",
    "azure-mgmt-resourcegraph>=8.0.0",
    "azure-mgmt-security>=7.0.0",
    "dotenv>=0.9.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "generativeai>=0.0.1",
    "google-generativeai>=0.8.5",
    "groq>=0.33.0",
    "openai>=2.6.1",
    "requests>=2.32.5",
    "six>=1.17.0",
]
